Fixed movement axis calculations to be based on hit normal and world up vector
Before it was keeping a reference to the surface mesh then grabbing the up and forward vectors from that.
But if I attached to the mesh from anywhere that wasn't the front of the mesh, the slime controls were not as you'd expect.
This has been fixed by calculating desired up and forward vectors on surface change.